[chore] merging some code
