+++ conversational_agent/src/sendEmail.ts
@@ -2,28 +2,81 @@ import dotenv from 'dotenv';
 import { JSDOM } from 'jsdom';
 import DOMPurify from 'dompurify';
 import { ServerClient } from 'postmark';
+import { getUserPreferences } from './getUserPreferences';
 
 // Load environment variables from the .env file
 dotenv.config({ path: './../.env' });
 
-export const sendMail = async (to: string, subject: string, body: string) => {
+export const sendMail = async (body: any) => {
     // Use the Postmark API key from environment variables
     const client = new ServerClient(process.env.POSTMARK_API_KEY || '');
 
     try {
+        const formattedBody = await buildResponse(body);
         // Send an email
         const result = await client.sendEmail({
             From: process.env.DEFAULT_POSTMARK_MAIL || '', // Replace with a verified email
-            To: to,
-            Subject: 'Re: ' + subject,
-            HtmlBody: formatHtmlBody(body),
-            TextBody: formatTextBody(body),
+            To: body['From'],
+            Subject: 'Re: ' + body['Subject'],
+            HtmlBody: formatHtmlBody(formattedBody),
+            TextBody: formatTextBody(formattedBody),
             MessageStream: 'outbound',
         });
         console.error('E-Mail sent successfully : ', result);
     } catch (error) {
         console.error('Error when trying to send the E-Mail :', error);
     }
+};
+
+const buildResponse = async (body: any) => {
+    // Generate a response from AI based on the received email text
+    const aiResponse = await getUserPreferences(body['TextBody']);
+
+    const window = new JSDOM('').window;
+    const purify = DOMPurify(window);
+    const cleanBodyFrom = purify.sanitize(body['From']);
+    const cleanBodyDate = purify.sanitize(body['Date']);
+    const cleanBodyTo = purify.sanitize(body['To']);
+    const cleanBodySubject = purify.sanitize(body['Subject']);
+    const cleanBodyTextBody = purify.sanitize(body['TextBody']);
+    const cleanThemes = purify.sanitize(
+        aiResponse?.themes.length == 1 ? 'theme' : 'themes'
+    );
+
+    const emailMetadata = `
+        
+        -------- Previous Message --------
+        
+        From: ${cleanBodyFrom}
+        
+        Sent: ${cleanBodyDate}
+        
+        To: ${cleanBodyTo}
+        
+        Subject: ${cleanBodySubject}
+        
+        ${cleanBodyTextBody}
+    
+    `;
+
+    if (!aiResponse?.themes?.length && !aiResponse?.unwanted_themes?.length) {
+        return `Sorry, we didn't find any preferences in your E-Mail. ${emailMetadata}`;
+    }
+    let textThemes = '';
+    if (aiResponse?.themes?.length) {
+        textThemes += `The following ${cleanThemes} have been added to your next newsletters :
+- ${aiResponse?.themes.join('\n  - ')}`;
+    }
+
+    let textUnwantedThemes = '';
+    if (aiResponse?.unwanted_themes?.length) {
+        textUnwantedThemes += `\nYou will no longer be annoyed with the following ${cleanThemes} :
+- ${aiResponse?.unwanted_themes.join('\n  - ')}`;
+    }
+    return `Hello!
+${textThemes}
+${textUnwantedThemes}
+    ${emailMetadata}`;
 };
 
 /**
