+++ conversational_agent/src/server.ts
@@ -1,6 +1,5 @@
 import express, { Request, Response } from 'express';
 import dotenv from 'dotenv';
-import { buildResponse } from './buildEmail';
 import { sendMail } from './sendEmail';
 import { JSDOM } from 'jsdom';
 import DOMPurify from 'dompurify';
@@ -36,10 +35,8 @@ app.post('/webhook', async (req: Request, res: Response) => {
 ${cleanBodyTextBody}`
     );
 
-    const response = await buildResponse(body);
-
     // Send a reply email with the content generated by OpenAI
-    await sendMail(body['From'], body['Subject'], response);
+    await sendMail(body);
 });
 
 // Start the server
